test(sauce): factorize the sauce configuration
