+++ karma-microtasks-sauce.conf.js
@@ -1,6 +1,4 @@
-// Karma configuration
-
 module.exports = function (config) {
-  require('./karma-browserify.conf')(config);
+  require('./karma-microtasks.conf')(config);
   require('./sauce.conf')(config);
 };
