+++ package.json
@@ -31,6 +31,7 @@
     "gulp-uglify": "^1.2.0",
     "jasmine-core": "^2.2.0",
     "karma": "^0.12.31",
+    "karma-browserify": "^4.1.2",
     "karma-chrome-launcher": "^0.1.7",
     "karma-firefox-launcher": "^0.1.4",
     "karma-jasmine": "^0.3.5",
