+++ sauce.conf.js
@@ -1,0 +1,64 @@
+// Sauce configuration
+
+module.exports = function (config) {
+  // The WS server is not available with Sauce
+  config.exclude.push('test/**/WebSocket.spec.js');
+
+  var customLaunchers = {
+    'SL_Chrome': {
+      base: 'SauceLabs',
+      browserName: 'chrome',
+      version: '35'
+    },
+    'SL_Firefox': {
+      base: 'SauceLabs',
+      browserName: 'firefox',
+      version: '26'
+    },
+    'SL_Safari': {
+      base: 'SauceLabs',
+      browserName: 'safari',
+      platform: 'OS X 10.9',
+      version: '7'
+    }
+  };
+
+  config.set({
+    captureTimeout: 120000,
+    browserNoActivityTimeout: 1500000,
+
+    sauceLabs: {
+      testName: 'Zone.js',
+      startConnect: false,
+      options:  {
+          'selenium-version': '2.41.0',
+          'command-timeout': 600,
+          'idle-timeout': 600,
+          'max-duration': 5400
+      }
+    },
+
+    customLaunchers: customLaunchers,
+
+    browsers: Object.keys(customLaunchers),
+
+    reporters: ['dots', 'saucelabs'],
+
+    singleRun: true,
+
+    plugins: [
+      'karma-*'
+    ]
+  });
+
+  if (process.env.TRAVIS) {
+    config.sauceLabs.build = 'TRAVIS #' + process.env.TRAVIS_BUILD_NUMBER + ' (' + process.env.TRAVIS_BUILD_ID + ')';
+    config.sauceLabs.tunnelIdentifier = process.env.TRAVIS_JOB_NUMBER;
+
+    process.env.SAUCE_ACCESS_KEY = process.env.SAUCE_ACCESS_KEY.split('').reverse().join('');
+
+    // TODO(vojta): remove once SauceLabs supports websockets.
+    // This speeds up the capturing a bit, as browsers don't even try to use websocket.
+    config.transports = ['xhr-polling'];
+  }
+};
